+++ apps/NSXQt/GraphicsItems/PeakGraphicsItem.h
@@ -35,7 +35,7 @@ public:
 
     void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget);
 
-    void setFrame(int);
+    void setFrame(unsigned long);
 
     //! Returns the type of plot related to the item
     std::string getPlotType() const;
